// src/enums/builtin-event-type.enum.ts
export enum BuiltinEventTypeEnum {
  ACTION = 'action',
  MESSAGE = 'message',
  PROGRAM = 'program',
  REPEAT_ACTION = 'repeat_action',
  STACK_CHANGE = 'stack_change',
  STACK_REMOVE = 'stack_remove',
  TARGET_SELECTION = 'target_selection',
  TARGET_SELECTED = 'target_selected',
  TRIGGER = 'trigger',
}